requirements.txt

streamlit
mysql-connector-python
pandasai
langchain_groq
python-dotenv
